requirements.txt
click==6.7
Flask==1.0.2
hapic==0.42
hapic-apispec==0.37.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
marshmallow==2.15.3
multidict==4.3.1
pkg-resources==0.0.0
python-vlc==3.0.102
PyYAML==3.12
Werkzeug==0.14.1
